+++ crates/jit-runtime/src/lib.rs
@@ -1,1 +1,8 @@
 pub use wasmtime_jit::*;
+pub mod mmap_code_memory;
+pub mod mmap_instantiate;
+pub mod profiling;
+pub mod unwind;
+
+pub use mmap_code_memory::MmapCodeMemory;
+pub use mmap_instantiate::CompiledModule;
