+++ crates/jit/Cargo.toml
@@ -15,7 +15,6 @@ wasmtime-environ = { workspace = true }
 wasmtime-jit-debug = { workspace = true, features = [
   "perf_jitdump",
 ], optional = true }
-wasmtime-runtime = { workspace = true }
 target-lexicon = { workspace = true }
 anyhow = { workspace = true }
 cfg-if = { workspace = true }
