+++ crates/jit/src/debug.rs
@@ -8,6 +8,7 @@ use object::{
 };
 use std::mem::size_of;
 
+/// TODO: doc
 pub fn create_gdbjit_image(
     mut bytes: Vec<u8>,
     code_region: (*const u8, usize),
