+++ crates/wasmtime/src/engine.rs
@@ -14,6 +14,7 @@ use wasmtime_cache::CacheConfig;
 use wasmtime_environ::obj;
 use wasmtime_environ::{FlagValue, ObjectKind};
 use wasmtime_jit_runtime::{profiling::ProfilingAgent, CodeMemory};
+use wasmtime_jit_runtime::{unwind::UnwindRegistration, MmapCodeMemory};
 use wasmtime_runtime::{CompiledModuleIdAllocator, InstanceAllocator, MmapVec};
 
 mod serialization;
@@ -607,7 +608,7 @@ impl Engine {
         &self,
         bytes: &[u8],
         expected: ObjectKind,
-    ) -> Result<Arc<CodeMemory>> {
+    ) -> Result<Arc<MmapCodeMemory>> {
         self.load_code(MmapVec::from_slice(bytes)?, expected)
     }
 
@@ -616,7 +617,7 @@ impl Engine {
         &self,
         path: &Path,
         expected: ObjectKind,
-    ) -> Result<Arc<CodeMemory>> {
+    ) -> Result<Arc<MmapCodeMemory>> {
         self.load_code(
             MmapVec::from_file(path).with_context(|| {
                 format!("failed to create file mapping for: {}", path.display())
@@ -625,9 +626,14 @@ impl Engine {
         )
     }
 
-    pub(crate) fn load_code(&self, mmap: MmapVec, expected: ObjectKind) -> Result<Arc<CodeMemory>> {
+    pub(crate) fn load_code(
+        &self,
+        mmap: MmapVec,
+        expected: ObjectKind,
+    ) -> Result<Arc<MmapCodeMemory>> {
         serialization::check_compatible(self, &mmap, expected)?;
-        let mut code = CodeMemory::new(mmap)?;
+        let mut code =
+            MmapCodeMemory::new(CodeMemory::new(mmap, UnwindRegistration::SECTION_NAME)?);
         code.publish()?;
         Ok(Arc::new(code))
     }
